E-waste-ML_model-api/requirements.txt
fastapi
uvicorn
ultralytics
requests
pillow
python-multipart
opencv-python-headless
onnxruntime